scripts/requirements.txt
pyyaml>=6.0
google-generativeai>=0.3.0
pandas>=2.0.0
tqdm>=4.65.0
click>=8.1.0
python-dotenv>=1.0.0